requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
google-api-python-client==2.149.0
google-auth==2.35.0
google-cloud-storage==2.18.2
google-cloud-firestore==2.16.1
